requirements.txt
streamlit
scikit-learn
chromadb
ollama
langchain-community
langchain-core
langchain-text-splitters
sentence-transformers
langchain-ollama
langchain-google-genai
google-generativeai
reportlab